restructuring talents
