+++ src/parser/priest/shadow/modules/talents/100/index.js
@@ -1,0 +1,5 @@
+import DarkAscension from './DarkAscension';
+
+export {
+  DarkAscension,
+};
