+++ src/parser/priest/shadow/modules/talents/45/TwistOfFate.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Analyzer from 'parser/core/Analyzer';
 
-import SPELLS from 'common/SPELLS';
+import SPELLS from 'common/SPELLS/index';
 import SpellIcon from 'common/SpellIcon';
 import { formatPercentage } from 'common/format';
 import SmallStatisticBox, { STATISTIC_ORDER } from 'interface/others/SmallStatisticBox';
