+++ src/parser/priest/shadow/modules/talents/45/index.js
@@ -1,0 +1,5 @@
+import TwistOfFate from './TwistOfFate';
+
+export {
+  TwistOfFate,
+};
