+++ src/parser/priest/shadow/modules/talents/90/VoidTorrent.js
@@ -1,12 +1,12 @@
 import React from 'react';
 
-import SPELLS from 'common/SPELLS';
+import SPELLS from 'common/SPELLS/index';
 import SpellLink from 'common/SpellLink';
 import SpellIcon from 'common/SpellIcon';
 import Analyzer from 'parser/core/Analyzer';
 import StatisticBox, { STATISTIC_ORDER } from 'interface/others/StatisticBox';
 
-import Voidform from './Voidform';
+import Voidform from '../../spells/Voidform';
 
 function formatSeconds(seconds) {
   return Math.round(seconds * 10) / 10;
