+++ src/parser/priest/shadow/modules/talents/90/index.js
@@ -1,0 +1,5 @@
+import VoidTorrent from './VoidTorrent';
+
+export {
+  VoidTorrent,
+};
