+++ src/parser/priest/shadow/modules/talents/index.js
@@ -1,0 +1,17 @@
+import * as talents_15 from './15';
+import * as talents_30 from './30';
+import * as talents_45 from './45';
+import * as talents_60 from './60';
+import * as talents_75 from './75';
+import * as talents_90 from './90';
+import * as talents_100 from './100';
+
+export default {
+  talents_15,
+  talents_30,
+  talents_45,
+  talents_60,
+  talents_75,
+  talents_90,
+  talents_100,
+};
